Enforce test file convention: merge feature-specific test files
- Added init_test.go to enforce convention: every *_test.go must have a corresponding *.go implementation
- Merged command_*_test.go files into command_test.go
- Merged map_clone_test.go into map_test.go
- Merged slice_cycle_test.go into slice_test.go
- Convention ensures test files remain paired with implementations and prevents orphaned tests
- Coverage: 86.9%
